+++ lib/src/feedback/components/intro_component.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:wiredash/src/common/translation/wiredash_translation.dart';
+import 'package:wiredash/src/common/translation/wiredash_localizations.dart';
 import 'package:wiredash/src/common/widgets/list_tile_button.dart';
 import 'package:wiredash/src/common/widgets/wiredash_icons.dart';
 import 'package:wiredash/src/feedback/feedback_model.dart';
@@ -22,8 +22,8 @@ class IntroComponent extends StatelessWidget {
             icon: WiredashIcons.bug,
             iconColor: const Color(0xff9c4db1),
             iconBackgroundColor: const Color(0xffffc4f0),
-            title: WiredashTranslation.of(context).feedbackModeBugTitle,
-            subtitle: WiredashTranslation.of(context).feedbackModeBugMsg,
+            title: WiredashLocalizations.of(context).feedbackModeBugTitle,
+            subtitle: WiredashLocalizations.of(context).feedbackModeBugMsg,
             onPressed: () => onModeSelectedCallback(FeedbackType.bug),
           ),
           const SizedBox(height: 12),
@@ -31,9 +31,10 @@ class IntroComponent extends StatelessWidget {
             icon: WiredashIcons.feature,
             iconColor: const Color(0xff007cbc),
             iconBackgroundColor: const Color(0xff2bd9fc),
-            title: WiredashTranslation.of(context).feedbackModeImprovementTitle,
+            title:
+                WiredashLocalizations.of(context).feedbackModeImprovementTitle,
             subtitle:
-                WiredashTranslation.of(context).feedbackModeImprovementMsg,
+                WiredashLocalizations.of(context).feedbackModeImprovementMsg,
             onPressed: () => onModeSelectedCallback(FeedbackType.improvement),
           ),
           const SizedBox(height: 12),
@@ -41,8 +42,8 @@ class IntroComponent extends StatelessWidget {
             icon: WiredashIcons.applause,
             iconColor: const Color(0xff00b779),
             iconBackgroundColor: const Color(0xffcdfbcb),
-            title: WiredashTranslation.of(context).feedbackModePraiseTitle,
-            subtitle: WiredashTranslation.of(context).feedbackModePraiseMsg,
+            title: WiredashLocalizations.of(context).feedbackModePraiseTitle,
+            subtitle: WiredashLocalizations.of(context).feedbackModePraiseMsg,
             onPressed: () => onModeSelectedCallback(FeedbackType.praise),
           ),
         ],